requirements.txt
python-telegram-bot[job-queue]
python-dotenv
gspread
oauth2client
pandas
google-generativeai